Custom validation

+++ testapiproject/Models/Student.cs
@@ -4,8 +4,8 @@ namespace testapiproject.Models
     {
         public int ID { get; set; }
         public string? Name { get; set; }
-        public string ?Email { get; set; }
-        public int Phone { get; set; }
+        public string? Email { get; set; }
+        public long Phone { get; set; }
 
     }
